requirements.txt
Fabric3
python-chi
python-swiftclient
python-heatclient
cryptography
croniter
paramiko
requests
ulid

# test
pytest
pytest-timeout
spur
